made ValidateLogin() a shared function and clear token on logout
