+++ angular/src/app/common.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CommonService } from './common.service';
+
+describe('CommonService', () => {
+  let service: CommonService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CommonService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
